Refactor for a better folder structure
