+++ FE/src/App.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom'
-import Login from './components/Login'
-import SignUp from './components/SignUp'
+import Login from './pages/auth/Login'
+import SignUp from './pages/auth/SignUp'
 
 function App() {
   return (
