+++ FE/src/components/Auth/shared/CornerDecorations.jsx
