+++ FE/src/components/Auth/shared/ErrorCorners.jsx
