+++ FE/src/components/ui/icons/EyeIcon.jsx
@@ -1,0 +1,10 @@
+import EyeIconSvg from '../../../assets/eye-icon.svg?react'
+
+/**
+ * Eye icon component - shows when password is hidden
+ */
+const EyeIcon = ({ className = 'w-5 h-5' }) => {
+  return <EyeIconSvg className={className} />
+}
+
+export default EyeIcon
