+++ FE/src/components/ui/icons/EyeOffIcon.jsx
@@ -1,0 +1,10 @@
+import EyeOffIconSvg from '../../../assets/eye-off-icon.svg?react'
+
+/**
+ * Eye-off icon component - shows when password is visible
+ */
+const EyeOffIcon = ({ className = 'w-5 h-5' }) => {
+  return <EyeOffIconSvg className={className} />
+}
+
+export default EyeOffIcon
